fix(hl7): add missing iomanip header and fix implicit conversion warnings
This commit fixes CI build failures on clang compilers (ubuntu-24.04 and
macos-14) caused by:

1. Missing <iomanip> header:
   - std::put_time requires <iomanip> header
   - Error occurred at line 632 when formatting datetime strings
   - Added #include <iomanip> to resolve the linker error

2. Implicit conversion warnings:
   - size_t to double conversions in update_processing_time() method
   - Added explicit static_cast<double>() for all numeric type conversions
   - Ensures clean compilation with -Wconversion flag

Tested locally with clang compiler to verify the fixes resolve the issues.

References: #15

+++ src/protocol/hl7/orm_handler.cpp
@@ -14,6 +14,7 @@
 
 #include <algorithm>
 #include <chrono>
+#include <iomanip>
 #include <sstream>
 
 namespace pacs::bridge::hl7 {
@@ -96,9 +97,11 @@ public:
 
     // Update statistics for processing time
     void update_processing_time(std::chrono::milliseconds elapsed) {
-        double total_time = stats_.avg_processing_ms * stats_.total_processed;
-        total_time += elapsed.count();
-        stats_.avg_processing_ms = total_time / (stats_.total_processed + 1);
+        double total_time =
+            stats_.avg_processing_ms * static_cast<double>(stats_.total_processed);
+        total_time += static_cast<double>(elapsed.count());
+        stats_.avg_processing_ms =
+            total_time / static_cast<double>(stats_.total_processed + 1);
     }
 };
